moved rules related endpoints to different file
