+++ server/endpoints.go
@@ -16,7 +16,12 @@ package server
 
 import (
 	"fmt"
+	"net/http"
+	"path/filepath"
 	"regexp"
+
+	"github.com/gorilla/mux"
+	"github.com/prometheus/client_golang/prometheus/promhttp"
 )
 
 const (
@@ -53,6 +58,49 @@ const (
 	// MetricsEndpoint returns prometheus metrics
 	MetricsEndpoint = "metrics"
 )
+
+func (server *HTTPServer) addDebugEndpointsToRouter(router *mux.Router) {
+	apiPrefix := server.Config.APIPrefix
+
+	router.HandleFunc(apiPrefix+OrganizationsEndpoint, server.listOfOrganizations).Methods(http.MethodGet)
+	router.HandleFunc(apiPrefix+DeleteOrganizationsEndpoint, server.deleteOrganizations).Methods(http.MethodDelete)
+	router.HandleFunc(apiPrefix+DeleteClustersEndpoint, server.deleteClusters).Methods(http.MethodDelete)
+	router.HandleFunc(apiPrefix+GetVoteOnRuleEndpoint, server.getVoteOnRule).Methods(http.MethodGet)
+	router.HandleFunc(apiPrefix+RuleEndpoint, server.createRule).Methods(http.MethodPost)
+	router.HandleFunc(apiPrefix+RuleErrorKeyEndpoint, server.createRuleErrorKey).Methods(http.MethodPost)
+	router.HandleFunc(apiPrefix+RuleEndpoint, server.deleteRule).Methods(http.MethodDelete)
+	router.HandleFunc(apiPrefix+RuleErrorKeyEndpoint, server.deleteRuleErrorKey).Methods(http.MethodDelete)
+
+	// endpoints for pprof - needed for profiling, ie. usually in debug mode
+	router.PathPrefix("/debug/pprof/").Handler(http.DefaultServeMux)
+}
+
+func (server *HTTPServer) addEndpointsToRouter(router *mux.Router) {
+	apiPrefix := server.Config.APIPrefix
+	openAPIURL := apiPrefix + filepath.Base(server.Config.APISpecFile)
+
+	// it is possible to use special REST API endpoints in debug mode
+	if server.Config.Debug {
+		server.addDebugEndpointsToRouter(router)
+	}
+
+	// common REST API endpoints
+	router.HandleFunc(apiPrefix+MainEndpoint, server.mainEndpoint).Methods(http.MethodGet)
+	router.HandleFunc(apiPrefix+ReportEndpoint, server.readReportForCluster).Methods(http.MethodGet, http.MethodOptions)
+	router.HandleFunc(apiPrefix+LikeRuleEndpoint, server.likeRule).Methods(http.MethodPut, http.MethodOptions)
+	router.HandleFunc(apiPrefix+DislikeRuleEndpoint, server.dislikeRule).Methods(http.MethodPut, http.MethodOptions)
+	router.HandleFunc(apiPrefix+ResetVoteOnRuleEndpoint, server.resetVoteOnRule).Methods(http.MethodPut, http.MethodOptions)
+	router.HandleFunc(apiPrefix+ClustersForOrganizationEndpoint, server.listOfClustersForOrganization).Methods(http.MethodGet)
+	router.HandleFunc(apiPrefix+DisableRuleForClusterEndpoint, server.disableRuleForCluster).Methods(http.MethodPut, http.MethodOptions)
+	router.HandleFunc(apiPrefix+EnableRuleForClusterEndpoint, server.enableRuleForCluster).Methods(http.MethodPut, http.MethodOptions)
+	router.HandleFunc(apiPrefix+RuleGroupsEndpoint, server.getRuleGroups).Methods(http.MethodGet, http.MethodOptions)
+
+	// Prometheus metrics
+	router.Handle(apiPrefix+MetricsEndpoint, promhttp.Handler()).Methods(http.MethodGet)
+
+	// OpenAPI specs
+	router.HandleFunc(openAPIURL, server.serveAPISpecFile).Methods(http.MethodGet)
+}
 
 // MakeURLToEndpoint creates URL to endpoint, use constants from file endpoints.go
 func MakeURLToEndpoint(apiPrefix, endpoint string, args ...interface{}) string {
