+++ server/server.go
@@ -44,10 +44,9 @@ package server
 
 import (
 	"context"
-	"encoding/json"
-	"io"
 	"net/http"
 	"net/url"
+
 	// we just have to import this package in order to expose pprof interface in debug mode
 	// disable "G108 (CWE-): Profiling endpoint is automatically exposed on /debug/pprof"
 	// #nosec G108
@@ -56,7 +55,6 @@ import (
 
 	"github.com/RedHatInsights/insights-operator-utils/responses"
 	"github.com/gorilla/mux"
-	"github.com/prometheus/client_golang/prometheus/promhttp"
 	"github.com/rs/zerolog/log"
 
 	"github.com/RedHatInsights/insights-results-aggregator/storage"
@@ -116,54 +114,6 @@ func (server *HTTPServer) listOfClustersForOrganization(writer http.ResponseWrit
 	if err != nil {
 		log.Error().Err(err).Msg(responseDataError)
 	}
-}
-
-func getTotalRuleCount(reportRules types.ReportRules) int {
-	totalCount := len(reportRules.HitRules) +
-		len(reportRules.SkippedRules) +
-		len(reportRules.PassedRules)
-	return totalCount
-}
-
-// getContentForRules returns the hit rules from the report, as well as total count of all rules (skipped, ..)
-func (server *HTTPServer) getContentForRules(
-	writer http.ResponseWriter,
-	report types.ClusterReport,
-) ([]types.RuleContentResponse, int, error) {
-	var reportRules types.ReportRules
-
-	err := json.Unmarshal([]byte(report), &reportRules)
-	if err != nil {
-		log.Error().Err(err).Msg("Unable to parse cluster report")
-		handleServerError(writer, err)
-		return nil, 0, err
-	}
-
-	totalRules := getTotalRuleCount(reportRules)
-
-	hitRules, err := server.Storage.GetContentForRules(reportRules)
-	if err != nil {
-		log.Error().Err(err).Msg("Unable to retrieve rules content from database")
-		handleServerError(writer, err)
-		return nil, 0, err
-	}
-
-	return hitRules, totalRules, nil
-}
-
-// getUserVoteForRules returns user votes for defined list of report's IDs
-func (server *HTTPServer) getUserVoteForRules(
-	feedbacks map[types.RuleID]types.UserVote,
-	rulesContent []types.RuleContentResponse,
-) []types.RuleContentResponse {
-	for i := range rulesContent {
-		if vote, found := feedbacks[types.RuleID(rulesContent[i].RuleModule)]; found {
-			rulesContent[i].UserVote = vote
-		} else {
-			rulesContent[i].UserVote = types.UserVoteNone
-		}
-	}
-	return rulesContent
 }
 
 func (server *HTTPServer) readReportForCluster(writer http.ResponseWriter, request *http.Request) {
@@ -230,21 +180,6 @@ func (server *HTTPServer) readReportForCluster(writer http.ResponseWriter, reque
 	}
 }
 
-// likeRule likes the rule for current user
-func (server *HTTPServer) likeRule(writer http.ResponseWriter, request *http.Request) {
-	server.voteOnRule(writer, request, types.UserVoteLike)
-}
-
-// dislikeRule dislikes the rule for current user
-func (server *HTTPServer) dislikeRule(writer http.ResponseWriter, request *http.Request) {
-	server.voteOnRule(writer, request, types.UserVoteDislike)
-}
-
-// resetVoteOnRule resets vote for the rule for current user
-func (server *HTTPServer) resetVoteOnRule(writer http.ResponseWriter, request *http.Request) {
-	server.voteOnRule(writer, request, types.UserVoteNone)
-}
-
 // checkUserClusterPermissions retrieves organization ID by checking the owner of cluster ID, checks if it matches the one from request
 func (server *HTTPServer) checkUserClusterPermissions(writer http.ResponseWriter, request *http.Request, clusterID types.ClusterName) error {
 	if server.Config.Auth {
@@ -261,93 +196,6 @@ func (server *HTTPServer) checkUserClusterPermissions(writer http.ResponseWriter
 		}
 	}
 	return nil
-}
-
-func (server *HTTPServer) voteOnRule(writer http.ResponseWriter, request *http.Request, userVote types.UserVote) {
-	clusterID, ruleID, userID, err := server.readClusterRuleUserParams(writer, request)
-	if err != nil {
-		// everything has been handled already
-		return
-	}
-
-	err = server.checkUserClusterPermissions(writer, request, clusterID)
-	if err != nil {
-		// everything has been handled already
-		return
-	}
-
-	err = server.Storage.VoteOnRule(clusterID, ruleID, userID, userVote)
-	if err != nil {
-		handleServerError(writer, err)
-		return
-	}
-
-	err = responses.SendOK(writer, responses.BuildOkResponse())
-	if err != nil {
-		log.Error().Err(err).Msg(responseDataError)
-	}
-}
-
-func (server *HTTPServer) getVoteOnRule(writer http.ResponseWriter, request *http.Request) {
-	clusterID, ruleID, userID, err := server.readClusterRuleUserParams(writer, request)
-	if err != nil {
-		// everything has been handled already
-		return
-	}
-
-	err = server.checkUserClusterPermissions(writer, request, clusterID)
-	if err != nil {
-		// everything has been handled already
-		return
-	}
-
-	userFeedbackOnRule, err := server.Storage.GetUserFeedbackOnRule(clusterID, ruleID, userID)
-	if err != nil {
-		handleServerError(writer, err)
-		return
-	}
-
-	err = responses.SendOK(writer, responses.BuildOkResponseWithData("vote", userFeedbackOnRule.UserVote))
-	if err != nil {
-		log.Error().Err(err).Msg(responseDataError)
-	}
-}
-
-// disableRuleForCluster disables a rule for specified cluster, excluding it from reports
-func (server *HTTPServer) disableRuleForCluster(writer http.ResponseWriter, request *http.Request) {
-	server.toggleRuleForCluster(writer, request, storage.RuleToggleDisable)
-}
-
-// enableRuleForCluster enables a previously disabled rule, showing it on reports again
-func (server *HTTPServer) enableRuleForCluster(writer http.ResponseWriter, request *http.Request) {
-	server.toggleRuleForCluster(writer, request, storage.RuleToggleEnable)
-}
-
-// toggleRuleForCluster contains shared functionality for enable/disable
-func (server *HTTPServer) toggleRuleForCluster(writer http.ResponseWriter, request *http.Request, toggleRule storage.RuleToggle) {
-	clusterID, ruleID, userID, err := server.readClusterRuleUserParams(writer, request)
-	if err != nil {
-		// everything has been handled already
-		return
-	}
-
-	err = server.checkUserClusterPermissions(writer, request, clusterID)
-	if err != nil {
-		// everything has been handled already
-		return
-	}
-
-	err = server.Storage.ToggleRuleForCluster(clusterID, ruleID, userID, toggleRule)
-	if err != nil {
-		log.Error().Err(err).Msg("Unable to toggle rule for selected cluster")
-		handleServerError(writer, err)
-		return
-	}
-
-	err = responses.SendOK(writer, responses.BuildOkResponse())
-	if err != nil {
-		log.Error().Err(err).Msg(responseDataError)
-	}
 }
 
 // getRuleGroups serves as a proxy to the insights-content-service redirecting the request
@@ -375,159 +223,6 @@ func (server *HTTPServer) getRuleGroups(writer http.ResponseWriter, request *htt
 	}
 
 	http.Redirect(writer, request, contentServiceURL.String()+RuleGroupsEndpoint, 302)
-}
-
-// deleteRuleForClusterToggle is debug endpoint for deleting the record in the rule_cluster_toggle table
-func (server *HTTPServer) deleteFromRuleClusterToggle(writer http.ResponseWriter, request *http.Request) {
-	clusterID, ruleID, userID, err := server.readClusterRuleUserParams(writer, request)
-	if err != nil {
-		// everything has been handled already
-		return
-	}
-
-	err = server.checkUserClusterPermissions(writer, request, clusterID)
-	if err != nil {
-		// everything has been handled already
-		return
-	}
-
-	err = server.Storage.DeleteFromRuleClusterToggle(clusterID, ruleID, userID)
-	if err != nil {
-		log.Error().Err(err).Msg("Unable to delete from rule_cluster_toggle")
-		handleServerError(writer, err)
-		return
-	}
-
-	err = responses.SendOK(writer, responses.BuildOkResponse())
-	if err != nil {
-		log.Error().Err(err).Msg(responseDataError)
-	}
-}
-
-func (server *HTTPServer) createRule(writer http.ResponseWriter, request *http.Request) {
-	ruleID, err := readRuleID(writer, request)
-	if err != nil {
-		// everything has been handled already
-		return
-	}
-
-	var rule types.Rule
-
-	err = json.NewDecoder(request.Body).Decode(&rule)
-	if err != nil {
-		if err == io.EOF {
-			err = &NoBodyError{}
-		}
-		handleServerError(writer, err)
-		return
-	}
-
-	rule.Module = ruleID
-
-	err = server.Storage.CreateRule(rule)
-	if err != nil {
-		handleServerError(writer, err)
-		return
-	}
-
-	err = responses.SendOK(writer, responses.BuildOkResponseWithData(
-		"rule", rule,
-	))
-	if err != nil {
-		log.Error().Err(err).Msg(responseDataError)
-	}
-}
-
-func (server *HTTPServer) deleteRule(writer http.ResponseWriter, request *http.Request) {
-	ruleID, err := readRuleID(writer, request)
-	if err != nil {
-		// everything has been handled already
-		return
-	}
-
-	err = server.Storage.DeleteRule(ruleID)
-	if err != nil {
-		handleServerError(writer, err)
-		return
-	}
-
-	err = responses.SendOK(writer, responses.BuildOkResponse())
-	if err != nil {
-		log.Error().Err(err).Msg(responseDataError)
-	}
-}
-
-func (server *HTTPServer) createRuleErrorKey(writer http.ResponseWriter, request *http.Request) {
-	ruleID, err := readRuleID(writer, request)
-	if err != nil {
-		// everything has been handled already
-		return
-	}
-
-	// it's gonna raise an error if rule does not exist
-	_, err = server.Storage.GetRuleByID(ruleID)
-	if err != nil {
-		handleServerError(writer, err)
-		return
-	}
-
-	errorKey, err := readErrorKey(writer, request)
-	if err != nil {
-		// everything has been handled already
-		return
-	}
-
-	var ruleErrorKey types.RuleErrorKey
-
-	err = json.NewDecoder(request.Body).Decode(&ruleErrorKey)
-	if err != nil {
-		if err == io.EOF {
-			err = &NoBodyError{}
-		}
-		handleServerError(writer, err)
-		return
-	}
-
-	ruleErrorKey.RuleModule = ruleID
-	ruleErrorKey.ErrorKey = errorKey
-
-	err = server.Storage.CreateRuleErrorKey(ruleErrorKey)
-	if err != nil {
-		handleServerError(writer, err)
-		return
-	}
-
-	err = responses.SendOK(writer, responses.BuildOkResponseWithData(
-		"rule_error_key", ruleErrorKey,
-	))
-	if err != nil {
-		log.Error().Err(err).Msg(responseDataError)
-	}
-}
-
-func (server *HTTPServer) deleteRuleErrorKey(writer http.ResponseWriter, request *http.Request) {
-	ruleID, err := readRuleID(writer, request)
-	if err != nil {
-		// everything has been handled already
-		return
-	}
-
-	errorKey, err := readErrorKey(writer, request)
-	if err != nil {
-		// everything has been handled already
-		return
-	}
-
-	err = server.Storage.DeleteRuleErrorKey(ruleID, errorKey)
-	if err != nil {
-		handleServerError(writer, err)
-		return
-	}
-
-	err = responses.SendOK(writer, responses.BuildOkResponse())
-	if err != nil {
-		log.Error().Err(err).Msg(responseDataError)
-	}
 }
 
 // readUserID tries to retrieve user ID from request. If any error occurs, error response is send back to client.
@@ -657,49 +352,6 @@ func (server *HTTPServer) Initialize(address string) http.Handler {
 	server.addEndpointsToRouter(router)
 
 	return router
-}
-
-func (server *HTTPServer) addDebugEndpointsToRouter(router *mux.Router) {
-	apiPrefix := server.Config.APIPrefix
-
-	router.HandleFunc(apiPrefix+OrganizationsEndpoint, server.listOfOrganizations).Methods(http.MethodGet)
-	router.HandleFunc(apiPrefix+DeleteOrganizationsEndpoint, server.deleteOrganizations).Methods(http.MethodDelete)
-	router.HandleFunc(apiPrefix+DeleteClustersEndpoint, server.deleteClusters).Methods(http.MethodDelete)
-	router.HandleFunc(apiPrefix+GetVoteOnRuleEndpoint, server.getVoteOnRule).Methods(http.MethodGet)
-	router.HandleFunc(apiPrefix+RuleEndpoint, server.createRule).Methods(http.MethodPost)
-	router.HandleFunc(apiPrefix+RuleErrorKeyEndpoint, server.createRuleErrorKey).Methods(http.MethodPost)
-	router.HandleFunc(apiPrefix+RuleEndpoint, server.deleteRule).Methods(http.MethodDelete)
-	router.HandleFunc(apiPrefix+RuleErrorKeyEndpoint, server.deleteRuleErrorKey).Methods(http.MethodDelete)
-
-	// endpoints for pprof - needed for profiling, ie. usually in debug mode
-	router.PathPrefix("/debug/pprof/").Handler(http.DefaultServeMux)
-}
-
-func (server *HTTPServer) addEndpointsToRouter(router *mux.Router) {
-	apiPrefix := server.Config.APIPrefix
-	openAPIURL := apiPrefix + filepath.Base(server.Config.APISpecFile)
-
-	// it is possible to use special REST API endpoints in debug mode
-	if server.Config.Debug {
-		server.addDebugEndpointsToRouter(router)
-	}
-
-	// common REST API endpoints
-	router.HandleFunc(apiPrefix+MainEndpoint, server.mainEndpoint).Methods(http.MethodGet)
-	router.HandleFunc(apiPrefix+ReportEndpoint, server.readReportForCluster).Methods(http.MethodGet, http.MethodOptions)
-	router.HandleFunc(apiPrefix+LikeRuleEndpoint, server.likeRule).Methods(http.MethodPut, http.MethodOptions)
-	router.HandleFunc(apiPrefix+DislikeRuleEndpoint, server.dislikeRule).Methods(http.MethodPut, http.MethodOptions)
-	router.HandleFunc(apiPrefix+ResetVoteOnRuleEndpoint, server.resetVoteOnRule).Methods(http.MethodPut, http.MethodOptions)
-	router.HandleFunc(apiPrefix+ClustersForOrganizationEndpoint, server.listOfClustersForOrganization).Methods(http.MethodGet)
-	router.HandleFunc(apiPrefix+DisableRuleForClusterEndpoint, server.disableRuleForCluster).Methods(http.MethodPut, http.MethodOptions)
-	router.HandleFunc(apiPrefix+EnableRuleForClusterEndpoint, server.enableRuleForCluster).Methods(http.MethodPut, http.MethodOptions)
-	router.HandleFunc(apiPrefix+RuleGroupsEndpoint, server.getRuleGroups).Methods(http.MethodGet, http.MethodOptions)
-
-	// Prometheus metrics
-	router.Handle(apiPrefix+MetricsEndpoint, promhttp.Handler()).Methods(http.MethodGet)
-
-	// OpenAPI specs
-	router.HandleFunc(openAPIURL, server.serveAPISpecFile).Methods(http.MethodGet)
 }
 
 // Start starts server
